LPS-170426 Create ConditionContainer component and add new conditions to show it
