+++ modules/apps/object/object-web/types/src/main/resources/META-INF/resources/js/components/ObjectAction/tabs/ConditionContainer.d.ts
@@ -1,0 +1,33 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+/// <reference types="react" />
+
+import './ActionBuilder.scss';
+import {ActionError} from '..';
+interface ConditionContainerProps {
+	disabled?: boolean;
+	errors: ActionError;
+	setValues: (values: Partial<ObjectAction>) => void;
+	validateExpressionURL: string;
+	values: Partial<ObjectAction>;
+}
+export declare function ConditionContainer({
+	disabled,
+	errors,
+	setValues,
+	validateExpressionURL,
+	values,
+}: ConditionContainerProps): JSX.Element;
+export {};
